spark/requirements.txt
pyspark
psycopg2-binary==2.9.1